change upload v2 test case

+++ QiniuSDKTests/QNConcurrentResumeUploadTest.m
@@ -167,7 +167,7 @@
     NSArray *sizeArray = @[@5000, @8000, @10000, @20000];
     NSInteger timestamp = [[NSDate date] timeIntervalSince1970];
     for (NSNumber *size in sizeArray) {
-        NSString *key = [NSString stringWithFormat:@"concurrent_resume_switch_region_v2_%@k_%d", size, timestamp];
+        NSString *key = [NSString stringWithFormat:@"concurrent_resume_switch_region_v2_%@k_%ld", size, (long)timestamp];
         QNTempFile *tempFile = [QNTempFile createTempFileWithSize:[size intValue] * 1024 identifier:key];
         [self allFileTypeSwitchRegionTestWithFile:tempFile key:key config:config option:nil];
     }
@@ -200,7 +200,7 @@
     NSArray *sizeArray = @[@500, @1000, @3000, @4000, @5000, @8000, @10000, @20000];
     NSInteger timestamp = [[NSDate date] timeIntervalSince1970];
     for (NSNumber *size in sizeArray) {
-        NSString *key = [NSString stringWithFormat:@"concurrent_http_v2_%@k_%d", size, timestamp];
+        NSString *key = [NSString stringWithFormat:@"concurrent_http_v2_%@k_%ld", size, (long)timestamp];
         QNTempFile *tempFile = [QNTempFile createTempFileWithSize:[size intValue] * 1024 identifier:key];
         [self allFileTypeUploadAndAssertSuccessResult:tempFile key:key config:config option:nil];
     }
@@ -217,7 +217,7 @@
     NSArray *sizeArray = @[@500, @1000, @3000, @4000, @5000, @8000, @10000, @20000];
     NSInteger timestamp = [[NSDate date] timeIntervalSince1970];
     for (NSNumber *size in sizeArray) {
-        NSString *key = [NSString stringWithFormat:@"concurrent_https_v2_%@k_%d", size, timestamp];
+        NSString *key = [NSString stringWithFormat:@"concurrent_https_v2_%@k_%ld", size, (long)timestamp];
         QNTempFile *tempFile = [QNTempFile createTempFileWithSize:[size intValue] * 1024 identifier:key];
         [self allFileTypeUploadAndAssertSuccessResult:tempFile key:key config:config option:nil];
     }
@@ -238,7 +238,7 @@
     NSArray *sizeArray = @[@30000];
     NSInteger timestamp = [[NSDate date] timeIntervalSince1970];
     for (NSNumber *size in sizeArray) {
-        NSString *key = [NSString stringWithFormat:@"concurrent_reupload_v2_%@k_%d", size, timestamp];
+        NSString *key = [NSString stringWithFormat:@"concurrent_reupload_v2_%@k_%ld", size, (long)timestamp];
         QNTempFile *tempFile = [QNTempFile createTempFileWithSize:[size intValue] * 1024 identifier:key];
         [self allFileTypeResumeUploadTest:0.5 * 1024 * size.longLongValue tempFile:tempFile key:key config:config option:nil];
     }
@@ -253,7 +253,7 @@
     }];
     
     NSInteger timestamp = [[NSDate date] timeIntervalSince1970];
-    NSString *keyUp = [NSString stringWithFormat:@"concurrent_NoKey_v2_%dk_%d", 600, timestamp];
+    NSString *keyUp = [NSString stringWithFormat:@"concurrent_NoKey_v2_%dk_%ld", 600, (long)timestamp];
     QNTempFile *tempFile = [QNTempFile createTempFileWithSize:600 * 1024 identifier:keyUp];
     tempFile.canRemove = NO;
     [self allFileTypeUploadAndAssertSuccessResult:tempFile key:nil config:configHttp option:nil];
